+++ lib/src/config/v2/server/backends.rs
@@ -4,7 +4,6 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq, Default)]
 #[serde(default)] // allow missing fields and fill them with the `..Self::default()` in this struct
 pub struct BackendSettings {
-    #[serde(skip)] // skip as long as there are no values
     pub rusty: RustyBackendSettings,
     #[serde(skip)] // skip as long as there are no values
     pub mpv: MpvBackendSettings,
@@ -13,10 +12,17 @@ pub struct BackendSettings {
 }
 
 /// Settings specific to the `rusty` backend
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq, Default)]
+#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
 #[serde(default)] // allow missing fields and fill them with the `..Self::default()` in this struct
 pub struct RustyBackendSettings {
-    // None for now
+    /// Enable or disable `soundtouch`; only has a effect if `rusty-soundtouch` is compiled-in
+    pub soundtouch: bool,
+}
+
+impl Default for RustyBackendSettings {
+    fn default() -> Self {
+        Self { soundtouch: true }
+    }
 }
 
 /// Settings specific to the `mpv` backend